gaia/requirements.txt
pyyaml
python-telegram-bot
pyotp
